requirements.txt
numpy==1.23.4
python-tictactoe==0.0.7
